+++ src/prelude.rs
@@ -1,9 +1,0 @@
-// number traits.
-pub use crate::num::OrdUtils;
-pub use crate::float::F32Ext;
-
-// memory.
-pub use crate::alloc::{Alloc, GlobalAlloc};
-pub use crate::arena::Arena;
-pub use crate::vec::Vec;
-