+++ changelog.md
@@ -2,6 +2,7 @@
 - now bumping minor for features & changes, patch for fixes.
 - `trait Alloc` is now unsafe (`Clone` requirement).
 - breaking change: `Vec::from_{value, fn}` now take the `len` before the value/fn.
+- add `mem::Align{1, 2, 4, 8, 16, 32, 64}`.
 
 0.1.11:
 - add `UnwrapDebug`.