+++ changelog.md
@@ -3,6 +3,7 @@
 - `trait Alloc` is now unsafe (`Clone` requirement).
 - breaking change: `Vec::from_{value, fn}` now take the `len` before the value/fn.
 - add `mem::Align{1, 2, 4, 8, 16, 32, 64}`.
+- turn `Into` impls into `From` impls.
 
 0.1.11:
 - add `UnwrapDebug`.