+++ changelog.md
@@ -7,6 +7,7 @@
 - simd `INF` & `NAN` consts.
 - simd `Default` impls.
 - add `erase`, `inc`, `enclose` macros.
+- replace `Rc::cast` with `{from, into}_inner`.
 
 0.1.11:
 - add `UnwrapDebug`.